SSG-9133: Prevent the software-only Gateway from blocking when console output is going to a buffer that is not being read.  The Process Controller enables console output using com.l7tech.server.log.console=true but reads STDIN and STDOUT from the subprocess only until it starts answering pings from the process controller, at which point the PC stops monitoring its console output.
Extra log categories are no longer added when console output is enabled unless extra categories are also explicitly requested.  The build.sh idea target now adds the extra categories, so log messages should show when the Gateway is run from IntelliJ as before, but will no longer be written to STDOUT in software-only Gateways running in production.

Summary:  the fix was in two parts, each sufficient to avoid the bulk of the problem:

1. The Gateway Java process was changed to avoid sending logs to the console by default.

2. The Gateway startup script was changed to arrange for anything written to the console to be read and ignored, to avoid having the Gateway process hang.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@41014 219f1c6e-4a0e-43c8-96f2-6e749be13fd4