SSM-4731 Solution Kit Installer.  Pass remote host (via request knob) to restman so that it can 1) set policy revision and 2) audit events with the logged in Policy Manager user (e.g. admin). See SSG-8291 for more details.
Outstanding: the "completed" audit event (pre-installation check and installation) still contains the incorrect user info:
     Admin User Name     : localsystem
     Admin User ID       : principal:localsystem

The "started" audit event is working correctly.  Note both events work correctly using wsman.  Is appears that restman is flushing the AuditContext.  Maybe it's possible for restman to support an optional flag to not flush the AuditContext?

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@41308 219f1c6e-4a0e-43c8-96f2-6e749be13fd4