SSM-4888: Fix for unhandled exception when user tries to select an icon for an encapsulated assertion.
The ImageCache now catches all exceptions while trying to parse an image resource, instead of just catching IOException, since we have now seen that the new image file format parsers will throw random unchecked exceptions when asked to parse corrupt image files.

It may also be a good idea to fix the broken "anonymous.gif", if anyone can figure out what it is supposed to be (I can't view it in IntelliJ either), or just remove it if it isn't used.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@41449 219f1c6e-4a0e-43c8-96f2-6e749be13fd4