SSG-10428:  Added Base32 encoding and decoding support to the Encode/Decode assertion.
The Encode/Decode assertion now accepts raw byte arrays (such as selected by ${gateway.random.83.binary}) as input values for encoding and decoding.  When a transformation requires character input and a byte[] array is provided, the character encoding configured in the assertion is used to convert the byte array to characters.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@41997 219f1c6e-4a0e-43c8-96f2-6e749be13fd4