SSG-10476: CassandraConnectionManager no longer implements ApplicationListener.  As it wasn't actually using the events for anything the ApplicationEventProxy was not required.
The Software DB keystore now persists a new empty Software DB KeyStore as soon as it is created, rather than waiting to do so until the first time the KeyStore is mutated (which is when the Default SSL key is created, later on).  This saves multiple empty KeyStore instances from being created and thrown away during the first Gateway startup with an empty database.  This also has the side effect of making the Software DB key store's database access pattern match that of the nCipher key store, which will make Spring problems less likely to require nCipher to be enabled in order to be detected.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@42083 219f1c6e-4a0e-43c8-96f2-6e749be13fd4