SSM-5024: The PolicyBackedKeyValueStoreAssertion no longer forgets to set the entity type in its return value from getEntitiesUsed().
The policy manager authoring environment and policy validator are now aware of encapsulated assertion or policy backed method arguments and return values while editing or validating a policy that acts as the backing policy for an encapsulated assertion or policy backed method.  Variables passed in as arguments will now be available while editing assertion properties in the policy as though they were set before the policy started (which in a sense they are).  Spurious validator warnings are no longer produced when an argument variable is used for the first time in a policy.  Validator warnings are now produced when a needed output variable is not set by the policy, or when it is set to a value that is explicitly the wrong data type.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@42146 219f1c6e-4a0e-43c8-96f2-6e749be13fd4