Added lots of JavaDoc to PolicyBackedServiceRegistry.
It is no longer necessary to create a Policy Backed Service instance in the database before using the policy backed services framework to consume an interface implemented by a policy provided that the interface has exactly one method.  Instead, you can just specify the interface along with the Goid of the Policy that will provide the implementation of its single method using the new PolicyBackedServiceRegistry interface getImplementationProxyForSingleMethod().  It is not required that any PolicyBackedService implementation instance exist before using this method to generate a proxy bound to a single method.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@42538 219f1c6e-4a0e-43c8-96f2-6e749be13fd4