SSG-11643: The Gateway now touches a file named "started" in the node/default/var directory as soon as it finishes delivering the ReadyForMessages application event (after which the listen ports are open).  This can be used by scripts to cleanly detect when a Gateway has finished starting up.
Note that the file will still be created even if certain subsystems suffered non-fatal startup errors -- notably, if a second instance of the Gateway is started while the first one is already running, the started file will still be touched even if the second instance was unable to open listen ports because they were already being held by the earlier instance.  So, be careful of that one.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@43389 219f1c6e-4a0e-43c8-96f2-6e749be13fd4