This is a code commit for the feature "Crypto Replacement", which implemented the following stories and tasks.
1. [US213778] Bouncy Castle for Non-FIPs Encryption.

[TA454703] Fix TeamCity unit tests that are failing
In this task, we fixed the following failed unit tests
(1) For PemUtilsTest and PEMSshHostKeyProviderTest
Add an override method, getPreferredProvider in new CryptoComplyJceProviderEngine and specify the service "Cipher.RSA/ECB/NoPadding" to find a preferred provider

(2) For ServerSymmetricKeyEncryptionDecryptionAssertionTest,
In ServerSymmetricKeyEncryptionDecryptionAssertion, ensure TripleDES key with 8 bytes is always converted to 16 bytes, so the new version of this assertion be compatible with policies from previous Gateway versions.

(3) For XencGcmTest,
Add one more error message "Tag mismatch" to identify Bouncy Castle is used and produces such error message.

(4) For EcdsaSignatureMethodTest
Explicitly specify the provide as Bouncy Castle when creating EcdsaSignatureMethod

(5) For ServerAsymmetricKeyEncryptionDecryptionTest,
Add an override method, getRsaNoPaddingCipherName in new CryptoComplyJceProviderEngine, and return "RSA/ECB/NoPadding"

(6) For AuditRecordCompatibilityVerifier52Test,
Handle a failure in a method "verifyAuditRecordSignature" in the class "AuditRecordCompatibilityVerifier52".

[TA465876] switch RSA BSAFE to bouncy castle for non-fips
In this task, we create a new provider engine to choose a JCE provider depending on FIPS configuration.  By default, Bouncy Castle in Non-FIPS is set.

2. [US214165] Gateway can start with CryptoComply Java for FIPS encryption

[TA506299] Fix gateway start failure due to switching to use CCJ FIPS
In this task, we added the new CCJ library and still keep RSA Bsafe library.  We fixed the shared key description for CCJ FIPS by converting secret key if the provider is CCJ and the algorithm is PBEWithSHA1AndDESede.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@47384 219f1c6e-4a0e-43c8-96f2-6e749be13fd4