US245156: The Gateway can now optionally create a TrustedCert entity corresponding to the default SSL key, if one is created (or imported) on startup.  The trust parameters can be defined using the system property "com.l7tech.bootstrap.autoTrustSslKey".  The Gateway does not create a TrustedCert entry unless this system property is set and nonempty.  An example value of the system property is "trustAnchor,verifyHostname,TrustedFor.SSL,TrustedFor.SAML_ISSUER" which would cause the new TrustedCert entry to be created as a trust anchor, with hostname verification enabled and trust flags set for outbound TLS and for signing SAML tokens.
There is currently no way to customize the name of the new TrustedCert entity (which will always be "AutoTrustSslKey") nor its revocation check policy (which will always be USE_DEFAULT).  There is also no way currently to tell the Gateway to add an entry corresponding to a higher-level cert than the subject cert, if the Gateway was configured to import an entire cert chain on startup to use for the default SSL key.

git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@47404 219f1c6e-4a0e-43c8-96f2-6e749be13fd4