DE254874: An accepted issuers list is no longer included by default for inbound TLS connections to listen ports with TLS 1.1 or TLS 1.2 enabled.  (They already weren't included if only TLS 1.0 was enabled.)  This restores compatibility with TLS clients that would not present a client certificate unless either its issuer was included in the accepted issuers list from the server, or else if the server sent an empty accepted issuers list.
To force an accepted issuers list to be included, set the advanced listen port property "acceptedIssuers" to true.  To additionally force the accepted issuers list to include the Gateway's default SSL cert if it would otherwise have been completely empty, set the advanced listen port property "includeSelfCertIfAcceptedIssuersEmpty" to true.

As before, to force an empty accepted issuers list to be presented, regardless of any other settings on the listen port, set the advanced listen port property "noAcceptedIssuers" to true.


git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@47675 219f1c6e-4a0e-43c8-96f2-6e749be13fd4