US352812 - "Implement solution for improved thead pool usage for client certificate login" HttpClient jar internals when using a PoolingClientConnectionManager stamps each created HTTP connection with the Principal obtained from the SSLContext when the admin user logs in with Client Certificate. Subsequent requests would then not reuse that connection, or any others, as the ‘state’ associated with the connection does not match the incoming state. The word ‘state’ means the value of attribute 'http.user-token’ in library.
The solution is to set the attribute 'http.user-token’ on a HttpContext and pass this into the HttpClient execute method. This stops the library from trying to find state later and instead stamps the connection with the value supplied.
To enable this the protected method createHttpContext is made public by way of new sub class DefaultHttpClientWithHttpContext which now sits in between SecureHttpComponentsClient and DefaultHttpClient. This enables the execute method taking a context to be invoked and to still use a context with the same configured attributes as before.

Thought was given to the value to use for the attribute ‘http.user-token'. The expected value of the attribute 'http.user-token’ is normally a Principal or null however it supports any value as is of type Object. The current value of the session id is used as the state as it is unique to the session and is consistent across the users session. This avoid the need to make further changes to enable the passing of the current User from module console into code in the common module. This data passing is currently being achieved through careful usage of the thread local in SessionSupport, which provides sessionId. sessionId is a good proxy for a user from a connection point of view as it is unique to the user logged on and any existing connections are closed when a user logs on.