US398788 - Add new ClientManager and interface
The manager is responsible for handling the configuration of clients.
Currently it only manages one client and topic, but could be expanded to
handle multiple topics.

This was added to remove a lot of the configuration from the module
listener, and keep the client portion of the code as dumb as possible.

Client manages the connection, disconnects, subscribing and handling of messages.
Manager handles starting/stopping the connection and configuring th
client topics/settings.