DE351400 fail to remove cluster prop (#694)
DE351400 - ClusterProperty cast in one of the checks within the handleClusterPropertyChange method was throwing when EntityInvalidationEvent, EntityVersionChecker is triggered. This is normally caused by a db content change resulting from another cluster node doing the update/create/delete. 
When the casting throws, it cases the applicationEvent chain to fail as well since the ApplicationEventProxy is handling the subscriber events sequentially with no error handling. 

I updated the EventTrackerManager to keep track of the GOID of the cleanup and forceopenlist CWP and remove the need of the clusterProperty class cast. 

I also added error handling for the ApplicationEventProxy in the event that when subscriber(s) throw an error the event chain can continue.