TA1056708 Policy cache performance improvements (#393)
* https://rally1.rallydev.com/#/40967354777d/detail/userstory/140192381516 - "Spike: investigate how to improve fragment cache performance”
POC of performance improvement by avoiding recompilation of policy within a single ‘batch’ of Entity Invalidation Events caused by the saving of a Policy.
Within Optum database, this change reduces the cost of updating the ‘Validate Required Header v2’ policy consistently from 87 seconds to 42 seconds on my Mac book pro.

The unit of operation in the DispatchingTransactionSynchronization is all entities modified during the last db transaction. When a policy is saved this includes every PolicyVersion and the Policy.
Policy’s are immutable in the sense that when ever they are updated a new version is created. Within a batch, if a Policy is recompiled (into a Server Policy) there is no need found yet that requires the recompilation of the Policy again.
As hibernate reports saves before updates (flushDirty) the PolicyVersion event is always produced first on a single node cluster in response to a Policy save, even though the Policy entity is updated before the new Policy Version is created. For this reason the optimization is only tested for this case from my local, the case where entity changes from another node are picked up is not tested. In this case the Policy entity would likely be processed first as they EntityManager is listed before the Policy Version entity manager in spring.
This is a POC branch and this code will be further improved

* https://rally1.rallydev.com/#/40967354777d/detail/userstory/140192381516 - "Spike: investigate how to improve fragment cache performance”
Added EntitiesProcessedInBatch as a Spring bean to manage tracking of updated entities during a batch processing updated entities.
This class managed whether this feature is on or not via boolean system property com.l7tech.server.avoid.policy.recompile, which is false by default.
Wired in new bean into PolicyCache and EntityVersionChecker
Updated existing tests for new dependency on EntitiesProcessedInBatch.

* https://rally1.rallydev.com/#/40967354777d/detail/userstory/140192381516 - "Spike: investigate how to improve fragment cache performance”
Turn log back down to fine

* https://rally1.rallydev.com/#/40967354777d/detail/userstory/140192381516 - "Spike: investigate how to improve fragment cache performance”
Update test for new constructor

* https://rally1.rallydev.com/#/40967354777d/detail/userstory/140192381516 - "Spike: investigate how to improve fragment cache performance”
Updated test application context for changes to beans

* some refactoring to get a easy understanding code and reduce sonarlint issues

* fix test

* Test fixes after code refactoring

* re-adding exception due to unit test

* refactoring: splitting code into methods and modernizing language usage

* some more refactoring

* Removing system property need and old behavior

* fix test config file

* fix unit test after rebasing

* Fix test versioning policy

* testing if policies are updated only required times

* effective testing for the number of policy compilation

* minor test change

* addressing code review comments