DE361031: Productize MC fixes for 9.4 (#963)
* DE361031: Productize MC fixes for 9.3 (#953)

* DE361031: Added com.l7tech.ncipher.position property to set the nCipher JCE provider priority arbitrarily, which replaces the use of com.l7tech.ncipher.preference property that always sets it to the top position (i.e. highest priority).
The default value of the com.l7tech.ncipher.position property is 1, meaning that it will still be set to the top position, unless specified otherwise. Mastercard will be instructed to set the property with a value of '0', which will cause it to be added as the last provider, after all those defined in the java.security file, making it the lowest priority. This numbering is consistent with the JCE provider management interface which is 1-based. This was a conscious decision (over making the last position the default) so that, if required, we can forward port this change for general release in CRs without impacting existing nCipher users, while allowing Mastercard (who is the only known FIPS Level 3 user) to upgrade without having to manually remove the system property from all their environments in the process.
Handling of an existing provider position has changed - if there is no priority specified, the existing position will be used; if there is a priority specified, the existing provider will be replaced. Prior to this change, the existing provider would only be used if the com.l7tech.ncipher.preference property was changed to a value that was NOT "highest" i.e. position 1. As the system property was later made so that it was programmatically set to "highest" if not already set to something else, this logic doesn't apply. As users should not be setting the java.security settings themselves, this change will have to be documented if made part of a general release.
Added logger to note the requested and resultant priorities given to the provider.

(cherry picked from commit cfde5f0)

* DE361031: Added explicit checks for KeyPairGenerator.RSA and KeyPairGenerator.EC in the getProviderFor() method in NcipherJceProviderEngine so they will be handled by nCipherKM regardless of how high actual priority it has. The provider engine should still be the item returned by JceProvider.getInstance(), even if it's not the highest priority provider. However, any service which is not explicitly stated to be handled by an engine class will be passed on to see if it's supported by the highest priority JCE provider that supports it. Because the nCipherKM provider can now be set to not be the highest priority, the KeyPairGenerator service can be supported by a higher priority provider and the generation will not happen on the HSM. In FIPS Level 3, keys generated outside of the HSM can't be imported, so key generation can fail. The new explicit check avoids this scenario.

(cherry picked from commit cf94033)

* DE361031: Removing explicit support for KeyPairGenerator.EC service from NcipherJceProviderEngine as we've never supported it under FIPS, apparently. Lack of support for generating EC key pairs on the HSM is not a regression.

(cherry picked from commit e580273)

* DE361031: Added a verifyMode switch to WssProcessorAlgorithmFactory that will cause getSignatureMethod to return (for RSAwithSHA1 only) a SignatureMethod from the highest priority default provider which will later be returned to a verification-specific SignatureMethod pool. This allows the use of Signature implementations from default providers for improved performance, instead of using the nCipherKM provider implementation (when enabled), which is required for signing with private keys on the HSM, but slow, especially when using FIPS Level 3. Only the WssProcessorImpl class sets verifyMode to enable at this point, as that's the source of the contention seen in the Mastercard thread dumps.

(cherry picked from commit d2d3818)

* DE361031: Added a verifyMode switch enabling to XmlElementVerifier, SamlAssertionV1, SamlAssertionV2, and ServerProcessSamlAuthnRequestAssertion, as those seem equally susceptible to major contention, as per repro environment testing.

(cherry picked from commit 6b7418f)

* DE361031: Added import statement missing in 9.2 which is required for hotfix merge.

* DE361031: Add system property to enable verifyMode switch. Default is false.

* DE361031: If com.l7tech.ncipher.position is not specified, fall back to using com.l7tech.ncipher.preference.

* DE361031: Address code review comments.

* DE361031: These files had conflicts due to line separtor changes (LF vs CRLF) in develop. During merge conflict resolution, I introduced some errors. Fixed the compile errors.